LayoutTests/streams/readablestream-cancel-infinite-stream-expected.txt

ReadableStream cancel should fullfil promise when cancel callback went fine.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rs.state is "closed"
PASS cancelationFinished is false
PASS storage.length is > 0
PASS chunkLengthRight is true
PASS successfullyParsed is true

TEST COMPLETE
